python-crew-service/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
python-dotenv
supabase
langchain-openai
langchain-deepseek
requests
pytest
pytest-mock
requests-mock
